requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
pydantic
python-dotenv
langsmith


